chore: isolate clock type (dependency injection helper)
