+++ todo/clock.go
@@ -1,0 +1,17 @@
+package todo
+
+import (
+	"time"
+)
+
+// Clock interface for time operations, allowing dependency injection for testing.
+type Clock interface {
+	Now() time.Time
+}
+
+// realClock implements Clock using the real time.
+type realClock struct{}
+
+func (realClock) Now() time.Time {
+	return time.Now()
+}
